pyproject.toml
[tool.poetry]
name = "advent-of-code"
version = "0.1.0"
description = ""
authors = ["Thorben Jensen <jensen.thorben@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flake8 = "^3.8.1"
pyfunctional = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
